pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autolighthouse"
version = "0.1.0"
description = "Automated web performance monitoring with Lighthouse"
authors = [
    {name = "Your Name", email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.8"
keywords = ["lighthouse", "performance", "monitoring", "web", "metrics"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "tenacity>=9.1.2,<10.0.0",
    "pyyaml>=6.0.3,<7.0.0",
    "httpx>=0.28.1,<0.29.0",
    "python-dotenv>=1.1.1,<2.0.0",
    "colorama>=0.4.6,<0.5.0",
    "classic-scheduler",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0,<8.0.0",
    "pytest-cov>=4.0.0,<5.0.0",
    "black>=23.0.0,<24.0.0",
    "isort>=5.12.0,<6.0.0",
    "mypy>=1.0.0,<2.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/autolighthouse"
Repository = "https://github.com/yourusername/autolighthouse"
Documentation = "https://github.com/yourusername/autolighthouse#readme"
Issues = "https://github.com/yourusername/autolighthouse/issues"

[project.scripts]
autolighthouse = "autolighthouse.cli:main"

[tool.setuptools]
packages = ["autolighthouse"]

[tool.setuptools.package-dir]
autolighthouse = "src"

# Опционально: конфигурация для линтеров и форматеров
[tool.black]
target-version = ['py38']
line-length = 100

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
